Esg_api_dynamic_1/requirements.txt
Flask
psycopg2-binary
cryptography
PyJWT
flask-cors
